Finish the assembler
