+++ tools/assembler.ps
@@ -1,0 +1,2 @@
+docker run -v "${pwd}:/assembler/assembler_context" -w /assembler/assembler_context hsc_assembler -- "$Args"
+
