+++ tools/assembler.sh
@@ -1,9 +1,7 @@
 #!/bin/sh
 
 # TODO: Add a windows version of this script (powershell or cmd.exe).
+# TODO: Add install prompt
 set -eu
 
-echo "The docker image (which this uses) doesn't work right now. Once the assembler is in a working state, that will be fixed."
-exit 0
-
-docker run -v "$PWD":/assembler_context -w /assembler_context hsc_assembler -- "$@"
+docker run -v "$PWD":/assembler/assembler_context -w /assembler/assembler_context hsc_assembler -- "$@"
