+++ tools/sources/assembler/README.md
@@ -1,3 +1,11 @@
 Assembler for this project.
 
 This uses the black auto formatter for code formatting
+
+# Usage
+
+To use this, either use the assembler.sh script which requires docker or install this from pypi.
+
+To install the pypi version, run `pip3 install hsc_assembler`.
+
+
