--- tools/sources/assembler/assembler/__init__.py
@@ -1,6 +1,0 @@
-from importlib import metadata
-
-__version__ = metadata.version(__name__)
-__all__ = ["assemble", "assemble_file"]
-
-from .assembler import assemble, assemble_file
