+++ tools/sources/assembler/hsc_assembler/__init__.py
@@ -1,0 +1,3 @@
+from hsc_assembler.main import assemble, assemble_file
+
+__all__ = ["assemble", "assemble_file"]
