+++ tools/sources/assembler/hsc_assembler/errors.py
