+++ tools/sources/assembler/hsc_assembler/parseable.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import typing
+from hsc_assembler import asm_parser
 
 if typing.TYPE_CHECKING:
     PT = typing.TypeVar("PT", bound="Parseable")
@@ -16,9 +17,8 @@ class Parseable:
         tokens: typing.Union[typing.Sequence[Token], Token],
         line: int,
     ) -> PT:
-        import parser
 
         if not isinstance(tokens, typing.Sequence):
             tokens = [tokens]
         func_name = f"parse_{cls.__name__}"
-        return getattr(parser, func_name)(cls, tokens, line)
+        return getattr(asm_parser, func_name)(cls, tokens, line)
