+++ tools/sources/assembler/hsc_assembler/utilities.py
