+++ tools/sources/assembler/tests/test_assembler.py
@@ -1,0 +1,10 @@
+# TODO: Add tests
+
+# TODO: Fix import issues. This doesn't run right now
+import hsc_assembler
+
+
+def test_assembler():
+    with open("tests/test.asm") as f:
+        code = f.read()
+    hsc_assembler.assemble(code)
